backend/requirements.txt
numpy==1.24.3
pillow==11.2.1
opencv-python-headless==4.8.1.78 